Commit 4 refactor Controller to RestController and add support HATEOAS

+++ src/main/java/org/ost/springboot/services/UsersService.java
@@ -2,6 +2,7 @@ package org.ost.springboot.services;
 
 import org.ost.springboot.models.User;
 import org.ost.springboot.repositories.UsersRepository;
+import org.ost.springboot.utils.UserNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,24 +25,26 @@ public class UsersService {
 
     public User findById(int id) {
         Optional<User> foundUser = usersRepository.findById(id);
-        return foundUser.orElse(null);
+        return foundUser.orElseThrow(UserNotFoundException::new);
     }
 
     @Transactional //т.к. метод не только читает данные, указываем аннотация Transactional, кторая заменит аннотицию класса
-    public void save(User user) {
-        usersRepository.save(user);
+    public User save(User user) {
+        return usersRepository.save(user);
     }
 
     @Transactional
-    public void update(int id, User updateUser) {
+    public User update(int id, User updateUser) {
         Optional<User> foundUser = usersRepository.findById(id);
         foundUser.ifPresent(user -> updateUser.setCreateAt(user.getCreateAt()));
         updateUser.setId(id);
-        usersRepository.save(updateUser);
+        return usersRepository.save(updateUser);
     }
 
     @Transactional
     public void delete(int id) {
+        if (usersRepository.findById(id).isEmpty())
+            throw new UserNotFoundException();
         usersRepository.deleteById(id);
     }
